Changed exception management
Board is designed with a middle layer approach:
- We have protected functions named aux_* that directly interact
with the data structures.
- We also have public functions that interact with the aux_* ones
and use them to create the logic needed for each purpose.

The first implementation relied on he aux_* functions to throw
exceptions when something went wrong, but this approach is bad,
since it uses the exceptions as an internal communication channel.

Therefore, code have been rearchitectures so that aux_* functions
return error codes instead (like bools) and then the public ones
(that are going to be the user interacts with) will throw excetions
if needed.

Also, the exceptions have been extracted from Board and templated
so that they can be used by the children classes with their own
relevant data, such as Position structures.