Added DefBoard
DefBoard is a Board which can always gives back an element when
calling getElement(Position).

To get this effect, a general default element must be passed at
initialization, which will be always returned at worst case escenario.
Additionaly, an element can be added and set (both steps needed and
independent from the "normal element" proccess) as default in different
positions (one default only per position, but a default doesn't have a
limit of positions), so that it has more relevance
in that position than the general default.

This way, when calling getElement(Position), the order of preference:

Element in position > Default element in position > General default